FeatureProvider separated from FeatureExtractor. Complete integration with Driver
Splitted the functionality from feature_extractor in FeatureExtractor
and FeatureProvider. FeatureExtractor is only for the extracting job and
doesn't care about the data source.
Also saving and loading is now the work of FeatureProvider.

All features are saved into a single file now, instead of using 3
different.

SavedFeatureProvider as removed as it's redundant with the current
infrastructure.
Saved data segmentations and features produce noew reproducable data.